+++ src/main/kotlin/org/jetbrains/research/ml/codetracker/ui/panes/TaskChoosingPane.kt
@@ -10,7 +10,7 @@ import javafx.fxml.FXML
 import javafx.scene.control.Button
 import javafx.scene.control.ComboBox
 import org.jetbrains.research.ml.codetracker.Plugin
-import org.jetbrains.research.ml.codetracker.TaskFileHandler
+import org.jetbrains.research.ml.codetracker.tracking.TaskFileHandler
 import org.jetbrains.research.ml.codetracker.server.PluginServer
 import org.jetbrains.research.ml.codetracker.ui.panes.util.*
 import java.net.URL
